backend/requirements.txt
repoze.sendmail==4.1
pyramid==1.7
pyramid_tm==0.12
pyramid_debugtoolbar
pyramid_authstack==1.0.1
SQLAlchemy==1.0.12
alembic==0.8.6
webhelpers2==2.0
transaction==1.4.3
zope.sqlalchemy==0.7.6
pyramid_mailer==0.14.1
redis==2.10.5
redlock-py==1.0.8
pyramid_jinja2==2.6.2
psycopg2==2.6.1
wtforms==2.1
celery==3.1.23
formencode==1.3.0
psutil==2.1.2
ziggurat_foundations>=0.6.7
bcrypt==2.0.0
appenlight_client
markdown==2.5
colander==1.2
defusedxml==0.4.1
dogpile.cache==0.5.7
pyramid_redis_sessions==1.0.1
simplejson==3.8.2
waitress==0.9.0
gunicorn==19.4.5
requests==2.9.1
requests_oauthlib==0.6.1
gevent==1.1.1
gevent-websocket==0.9.5
pygments==2.1.3
lxml==3.6.0
paginate==0.5.4
paginate-sqlalchemy==0.2.0
pyelasticsearch==1.4
six==1.8.0
mock==1.0.1
itsdangerous==0.24
camplight==0.9.6
jira==0.41
python-dateutil==2.5.3
authomatic==0.1.0.post1
cryptography==1.2.3
webassets==0.11.1

